pkg/proxy: rename to pkg/server
